requirements.txt
# Telegram API
telethon==1.35.0
aiogram==3.4.1

# Async support
asyncio-mqtt==0.13.0
aiohttp==3.9.1
aiofiles==23.2.0

# Database
redis==5.0.1
aiosqlite==0.19.0

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2

# Security
cryptography==41.0.8

# Utils
python-dateutil==2.8.2
pytz==2023.3 